create: Move FormulaCreator to formula_creator.rb
